Only copy t2embed.dll if it doesn't already exist (#596)
I have run into a situation, where t2embed.dll already was in place
causing the navinstall to fail, this PR ensures that the file is only
copied if it doesn't already exist.

+++ generic/Run/130/navinstall.ps1
@@ -102,7 +102,9 @@ Copy-Item -Path (Join-Path $roleTailoredClientFolder "*office*.dll") -Destinatio
 # Due to dependencies from finsql.exe, we have to copy hlink.dll and ReportBuilder in place inside the container
 Copy-Item -Path (Join-Path $runPath 'Install\hlink.dll') -Destination (Join-Path $roleTailoredClientFolder 'hlink.dll')
 Copy-Item -Path (Join-Path $runPath 'Install\hlink.dll') -Destination (Join-Path $serviceTierFolder 'hlink.dll')
-Copy-Item -Path (Join-Path $runPath 'Install\t2embed.dll') -Destination "c:\windows\system32\t2embed.dll"
+if (!(Test-Path "c:\windows\system32\t2embed.dll")) {
+    Copy-Item -Path (Join-Path $runPath 'Install\t2embed.dll') -Destination "c:\windows\system32\t2embed.dll"
+}
 Copy-Item -Path (Join-Path $runPath 'Install\Microsoft.IdentityModel.dll') -Destination (Join-Path $serviceTierFolder 'Microsoft.IdentityModel.dll')
 
 $reportBuilderPath = "C:\Program Files (x86)\ReportBuilder"
